+++ maserol/predictAbundKa.py
@@ -10,7 +10,7 @@ from tqdm import tqdm
 from scipy.optimize import minimize
 from jax import value_and_grad, jit, jacfwd, jacrev
 from jax.config import config
-from scipy.stats import pearsonr
+
 from tensorly.decomposition import non_negative_parafac
 
 
@@ -40,6 +40,7 @@ def infer_Lbound(R_subj, R_Ag, Ka, L0=1e-9, KxStar=1e-12):
     pass the matrices generated above into polyfc, run through each receptor
     and ant x sub pair and store in matrix same size as flatten
     """
+    # TODO: make this capable of handling missing data
     Phisum = jnp.zeros((R_subj.shape[0], Ka.shape[0], R_Ag.shape[0]))
     Rtot = jnp.einsum("ij,kj->ijk", R_subj, R_Ag)
 
@@ -104,32 +105,3 @@ def optimize_lossfunc(cube, n_ab=1, maxiter=100):
 
     return opt.x[:, np.newaxis]
 
-
-def plot_correlation_heatmap(ax, RKa_opt, cube, rec_names, ant_names):
-    """
-    Uses optimal parameters from optimize_lossfunc to run the model
-    Generates prelim figures to compare experimental and model results
-    R_subj, R_Ag, Ka, L0=L0, KxStar=KxStar
-    """
-
-    R_subj, R_Ag, Ka = reshapeParams(RKa_opt, cube)
-    Lbound_model = infer_Lbound(R_subj, R_Ag, Ka, L0=1e-9, KxStar=1e-12)
-
-    coeff = np.zeros([cube.shape[1], cube.shape[2]])
-    for ii in range(cube.shape[1]):
-        for jj in range(cube.shape[2]):
-            coeff[ii, jj], _ = pearsonr(cube[:, ii, jj], Lbound_model[:, ii, jj])
-
-    print(coeff)
-    ax.imshow(coeff)
-
-    # Show all ticks and label them with the respective list entries
-    ax.set_xticks(np.arange(len(ant_names)))
-    ax.set_xticklabels(ant_names, rotation=45)
-    ax.set_yticks(np.arange(len(rec_names)))
-    ax.set_yticklabels(rec_names)
-
-    # Loop over data dimensions and create text annotations.
-    for i in range(len(ant_names)):
-        for j in range(len(rec_names)):
-            text = ax.text(i, j, round(coeff[j, i], 2), ha="center", va="center", color="w")
